// src/tests/generated/CastBallot.ts
/* tslint:disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: CastBallot
// ====================================================

export interface CastBallot {
  castBallot: boolean;
}

export interface CastBallotVariables {
  electionId: string;
  candidateIds: string[];
}